Consolidated select tests, got rid of queries file
